Can't work out for the life of me how to connect to this heinzinger...Serial connection is doing weird stuff I don't fully understand, maybe a simpler program might be better to start

+++ include/HeinzingerVoltageWindow.h
@@ -14,7 +14,7 @@ signals:
 	void set_voltage(uint voltage);
 	
 public slots:
-	void readbackVoltage(int voltage);
+	void readbackVoltage(qreal voltage);
 	
 private slots:
 	void changeVoltage();
